+++ index.js
@@ -3,7 +3,19 @@ module.exports = {
     'selector-pseudo-class-no-unknown': [
       true,
       {
-        ignorePseudoClasses: ['export', 'import', 'global', 'local'],
+        ignorePseudoClasses: [
+          'export',
+          'import',
+          'global',
+          'local',
+          'external',
+        ],
+      },
+    ],
+    'selector-type-no-unknown': [
+      true,
+      {
+        ignoreTypes: ['from'],
       },
     ],
     'property-no-unknown': [